+++ src-electron/core/world/operations/install.ts
@@ -1,0 +1,5 @@
+import { CustomMapData } from 'app/src-electron/schema/filedata';
+
+export function InstallCustomMap(map: CustomMapData) {
+  map;
+}
